Fix payload sender appending (#879)
* Append sender bytes to payload

* Send modified payload where it's supposed to be

* Cut the first byte from the sender

We want to restrict this to 32 bytes.  Cutting the first byte is how we
decided to do this because it has the least entropy

* Switch to appending hex address of sender

We want to be consistent, and sending the bytes of the contract address
is too much.  Instead, we're not hex encoding the address and sending
the bytes of that.  The same is being done for non-contract initiated
executions for consistency

* Zero pad sender bytes

We want to send 32 bytes of data for the sender address.  When the
address is less than this, we want to zero pad the front of it.

---------

Co-authored-by: Tyler Ruppert <{ID}+{username}@users.noreply.github.com>