Publish valset to all chains when returning from a software update (#942)
* Fixing error messages based on fmt formatting

* Publish valset to all chains when returning from a software update

Adding the ability to force publish a valset to all chains.  Then taking
this ability and applying it to when we come back from an upgrade.  We
are very often just above 2/3 previous consensus when we come back from
an upgrade, so we should publish the valset that has changed by that
much, in case any validators have trouble right after the upgrade and go
offline.

---------

Co-authored-by: Tyler Ruppert <{ID}+{username}@users.noreply.github.com>