Expose a query endpoint for latest published valset (#985)
* Expose a query endpoint for latest published valset

Pigeon currently gets this information by querying the compass contract.
The problem is, when there is a just-in-time valset update, pigeon will
end up using an old valset because the new valset hasn't been committed
yet.  We need the newest valset that may still be pending on the evm
chain.

Now, we can query paloma instead of the evm contract.  Paloma knows what
the latest published valset is, even if it hasn't yet been committed.

* Add tests

---------

Co-authored-by: Tyler Ruppert <{ID}+{username}@users.noreply.github.com>