feat!: #795 Transfer ERC20 ownership on compass update (#1002)
* chore: replace deprecated math API

* feat: yet another logging lib

* chore: improve logging during attestation

* feat!: #795 Transfer ERC20 ownership on compass update

Since compass is the owner of any ERC20 tokens deployed to target
chains, we need to make sure we transfer the ownership to the
new compass during the upgrade process.

This change alters the current upgrade procedure. Once the new
compass is reported as deployed, we send a new message to Pigeon,
instructing it to perform the transfer. Only after we get a
successful ownership transfer attestation will we change
Paloma to use the new compass contract.

Also contains a small wrapper for the default logging framework,
as I found no default reliable way of logging errors as well as
some housekeeping.

* chore: remove unused import

* fix: update test CI to go 1.20

* fix: use proper address interface

* ci: yaml is weird ...