fix: prevent TPK when not attesting faulty msg (#1099)
# Related Github tickets

- https://github.com/VolumeFi/paloma/issues/1425

# Background

This change adds two safe guards to prevent a mass jailing in case of a
faulty message:

a) messages that do not build consensus because neither `errorData` nor
`publicAccessData` has been added, meaning the message was never handled
in the first place. Nobody should be punished for not attesting what's
not attestable.
b) messages that less than 10% of the network attest are treated as
indications of faulty message responses which cannot be attested to even
when tried. Nobody will be punished for missing attestation on such a
message.

# Testing completed

- [x] test coverage exists or has been added/updated
- [x] tested in a private testnet

# Breaking changes

- [x] I have checked my code for breaking changes
- [x] If there are breaking changes, there is a supporting migration.