feat: lightnode sale (#1224)
# Related Github tickets

- Closes https://github.com/VolumeFi/paloma/issues/1866

# Background

This PR prepares Paloma for the light node client registration. It adds:
- Paloma light node feegranter governance vote - as the light nodes will
require feegrants, this vote determines who pays for the fees.
- Light node client registration transaction - to register a sold light
node and start vesting
- Manual light node client license transaction - allows to create light
node licenses ad-hoc, with the creator as the source of funds
- Queries for light node licenses and feegranter

NOTE: this requires an upgrade, as the paloma module did not yet have a
store. It is currently set to the 1.15.4 release, but needs be changed
before merging if this gets pushed to a later release.

# Testing completed

- [x] test coverage exists or has been added/updated
- [x] tested in a private testnet

# Breaking changes

- [x] I have checked my code for breaking changes
- [x] If there are breaking changes, there is a supporting migration.