Fix/pigeons attest rolledback txs (#1284)
# Related Github tickets

- https://github.com/VolumeFi/paloma/issues/2095

# Background

Paloma was attesting to pigeon relayed messages using only the
transaction, but not the receipt, meaning it couldn't actually attest
the transaction was successful. This (and a corresponding PR in pigeon)
makes paloma verify the transaction result from the transaction receipt
and only attest to successful transactions.

Validated with an SLC message in the private testnet using Gnosis, but
the result should be the same for all turnstone messages. This wasn't
replicable in Arbitrum, where messages fail immediately on pigeons.

Also in this PR is a fix for SLC and User Smart Contract Upload retries.
We need to clear the Fees from the message before a retry, otherwise we
get an Invalid Signature error.

# Testing completed

- [x] test coverage exists or has been added/updated
- [x] tested in a private testnet

# Breaking changes

- [x] I have checked my code for breaking changes
- [x] If there are breaking changes, there is a supporting migration.