feat: add pigeon metrics to telemetry endpoint (#1287)
# Related Github tickets

- Closes https://github.com/VolumeFi/paloma/issues/2046

# Background

This PR adds pigeon metrics to the telemetry endpoint, where they can be
consumed in either JSON or prometheus format.

To get the metrics, we need to enable telemetry on app.toml -
`telemetry.enabled = true`. If we want to enable the prometheus format,
we need to enable prometheus sink as well, by setting prometheus
retention to non-zero, e.g. `telemetry.prometheus-retention-time = 60`
to retain metrics for a minute.
Metrics will be available on the API, in the `/metrics` endpoint, e.g.
`http://localhost:1317/metrics?format=prometheus`
Pigeon metrics keys are prefixed with `pigeon_`.

# Testing completed

- [ ] test coverage exists or has been added/updated
- [x] tested in a private testnet

# Breaking changes

- [x] I have checked my code for breaking changes
- [x] If there are breaking changes, there is a supporting migration.