chore: update validator nonces in endblocker (#1292)
# Related Github tickets

- https://github.com/VolumeFi/paloma/issues/2160

# Background

The way it was before, we would need a claim to be observed so
validators would be updated. By doing it in the endblocker, we
validators are always up-to-date.

This should prevent a situation where the last claim is 19, the last
observed nonce is 18, but validators are stuck on nonce 17. They would
receive a list of block events with the block of nonce 19 (latest
unobserved) but would be unable to claim it, since skyway would expect
them to submit a claim for nonce 18 instead.

# Testing completed

- [ ] test coverage exists or has been added/updated
- [x] tested in a private testnet

# Breaking changes

- [x] I have checked my code for breaking changes
- [x] If there are breaking changes, there is a supporting migration.